embeddings_and_cosine_similarity/faiss-vdb/requirements.txt
faiss-cpu
numpy



# Sentence Transformers
sentence-transformers